change invalid media type to error
