+++ packages/shared/src/lib/core/nfts/enums/download-error-type.enum.ts
@@ -1,4 +1,5 @@
 export enum DownloadErrorType {
+    UnsupportedMediaType = 'unsupportedMediaType',
     NotMatchingFileTypes = 'notMatchingFileTypes',
     UnsupportedUrl = 'unsupportedUrl',
     NotReachable = 'notReachable',
