+++ packages/shared/src/lib/core/nfts/enums/download-warning-type.enum.ts
@@ -2,7 +2,6 @@ export enum DownloadWarningType {
     TooLargeFile = 'tooLargeFile',
     DownloadTooLong = 'downloadTooLong',
     DownloadNotAllowed = 'downloadNotAllowed',
-    UnsupportedMediaType = 'unsupportedMediaType',
     Generic = 'generic',
     ScamNft = 'scamNft',
 }
